+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade K8s dependencies ([#1149](https://github.com/giantswarm/kubectl-gs/pull/1149)).
+  - Upgrade giantswarm/k8sclient to 7.1.0.
+  - Upgrade sigs.k8s.io/cluster-api to v1.5.2.
+  - Migrate sigs.k8s.io/cluster-api-provider-aws to v2 (2.2.4).
+  - Upgrade sigs.k8s.io/cluster-api-provider-azure to v1.11.4.
+  - Upgrade sigs.k8s.io/controller-runtime to v0.16.3.
+  - Upgrade github.com/coreos/go-oidc/v3 to v3.6.0.
+  - Upgrade other dependencies.
+  - Replace capi.ClusterLabelName with capi.ClusterNameLabel.
+
 ## [2.45.0] - 2023-10-25
 
 ### Added